requirements.txt
streamlit==1.51.0
groq==1.1.2
python-dotenv==1.0.0
